[EN-657] services/topics: move add, update and delete topic into custom_topics module
